Pierwszy refaktor - podział całego algorytmu na 3 klasy
