+++ src/main/java/pl/zzpj/cryptography/des/algorithm/DES.java
@@ -4,19 +4,21 @@ import java.util.Arrays;
 
 import pl.zzpj.cryptography.des.utils.ArrayUtils;
 import pl.zzpj.cryptography.des.utils.BitJuggler;
+import pl.zzpj.cryptography.des.utils.MatrixUtils;
 
 public class DES {
   
-  private byte[][] subKeys;
+  private FFunction fFunction;
   
   public DES(byte[] key) {
-    this.subKeys = this.generateSubKeys(key);
+    this.fFunction = new FFunction(key);
   }
 
-  public byte[] enctrypt(byte[] src, boolean isDecrypt) {
+  public byte[] encrypt(byte[] src, boolean isDecrypt) {
     byte[][] srcIn2D = ArrayUtils.transformToTwoDimensionsArray(src);
     
     byte[][] tmp = new byte[srcIn2D.length][8];
+    
     for (int i = 0; i < tmp.length; i++) {
       tmp[i] = this.encrypt8ByteBlock(srcIn2D[i], isDecrypt);
     }
@@ -39,8 +41,9 @@ public class DES {
     return ArrayUtils.transformToOneDimensionArray(tmp);
   }
   
-  public byte[] encrypt8ByteBlock(byte[] block, boolean isDecrypt) {
-    byte[] permutedOrginalBlock = this.permute(block, DESPermutationTables.IP);
+  
+  private byte[] encrypt8ByteBlock(byte[] block, boolean isDecrypt) {
+    byte[] permutedOrginalBlock = MatrixUtils.permute(block, DESPermutationTables.IP);
     
     byte[] left = BitJuggler.getBits(permutedOrginalBlock, 0, 32);
     byte[] right = BitJuggler.getBits(permutedOrginalBlock, 32, 32);
@@ -49,9 +52,9 @@ public class DES {
       byte[] tmpR = right;
       
       if (isDecrypt) {
-        right = this.performFFunction(right, this.subKeys[15 - i]);
+        right = fFunction.perform(right, 15 - i);
       } else {
-        right = this.performFFunction(right, this.subKeys[i]);
+        right = fFunction.perform(right, i);
       }
       
       right = BitJuggler.xorArrays(left, right);
@@ -60,74 +63,10 @@ public class DES {
     }
     
     byte[] result = BitJuggler.concatBitSeries(right, 32, left, 32);    
-    result = this.permute(result, DESPermutationTables.FP);
+    result = MatrixUtils.permute(result, DESPermutationTables.FP);
     
     return result;
   }
   
-  private byte[] permute(byte[] input, byte[] permutationTable) {
-    byte[] result = new byte[(permutationTable.length - 1) / 8 + 1];
-    for (int i = 0; i < permutationTable.length; i++) {
-      int value = BitJuggler.getBit(input, permutationTable[i] - 1);
-      BitJuggler.setBit(result, i, value);
-    }
-    
-    return result;
-  }
-  
-  private byte[] useSBoxes(byte[] input) {
-    byte[] inputAfterSeparation = BitJuggler.separateBits(input, 6);      // length = 8
-    byte[] rowAndColumn = {0, 0}; // 00000000 00000000 // 0-7 row bits, 8-15 column bits
-    byte[] result = new byte[4];
-    
-    for (int i = 0, j = 0; i < inputAfterSeparation.length; i++, j += 4) {
-      byte[] currentByte = { inputAfterSeparation[i] };
-      
-      // Row
-      BitJuggler.setBit(rowAndColumn, 6, BitJuggler.getBit(currentByte, 0));
-      BitJuggler.setBit(rowAndColumn, 7, BitJuggler.getBit(currentByte, 5));
-      
-      // Column
-      BitJuggler.setBits(rowAndColumn, 12, currentByte, 1, 4);
-      
-      byte row = rowAndColumn[0];
-      byte column = rowAndColumn[1];
-      byte[] sBoxValue = { DESPermutationTables.S_BOXES[i][row][column] };
-      
-      BitJuggler.setBits(result, j, sBoxValue, 4, 4);
-      
-    }
-    
-    return result;
-  }
-  
-  private byte[] performFFunction(byte[] input, byte[] key) {
-    byte[] result;
-    result = this.permute(input, DESPermutationTables.E);
-    result = BitJuggler.xorArrays(result, key);
-    result = this.useSBoxes(result);
-    result = this.permute(result, DESPermutationTables.P);
-    
-    return result;
-  }
-  
-  public byte[][] generateSubKeys(byte[] key) {
-    byte[][] subKeys = new byte[16][];
-    byte[] permutedKey = this.permute(key, DESPermutationTables.PC1);
-    
-    byte[] C0 = BitJuggler.getBits(permutedKey, 0, 28);
-    byte[] D0 = BitJuggler.getBits(permutedKey, 28, 28);
-    
-    for (int i = 0; i < 16; i++) {
-      C0 = BitJuggler.rotateSelectedBitsLeft(C0, 28, DESPermutationTables.SUBKEY_ROTATIONS[i]);
-      D0 = BitJuggler.rotateSelectedBitsLeft(D0, 28, DESPermutationTables.SUBKEY_ROTATIONS[i]);
-      
-      byte[] C0AndD0Together = BitJuggler.concatBitSeries(C0, 28, D0, 28);
-      C0AndD0Together = this.permute(C0AndD0Together, DESPermutationTables.PC2);
-      subKeys[i] = C0AndD0Together; 
-    }
-    
-    return subKeys;
-  }
   
 }
